requirements.txt
# Requirements for the Gemini chatbot
requests
google-genai
pydantic
